requirements.txt
kivy>=2.1.0
openai==0.28
python-dotenv
